test: Refactor unit/integration/e2e tests
Change-type: patch
