+++ test/unit/core/cards/john-doe-no-type.json
